Implementing sync feature
